backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
boto3>=1.34
pillow>=10.0  # for image uploads

